+++ packages/graphiql/src/components/DocExplorer/SchemaDoc.tsx
@@ -28,6 +28,7 @@ export default function SchemaDoc({ schema, onClickType }: SchemaDocProps) {
       <MarkdownContent
         className="doc-type-description"
         markdown={
+          schema.description ||
           'A GraphQL schema provides a root type for each kind of operation.'
         }
       />